+++ src/harlequin/actions.py
@@ -250,6 +250,9 @@ HARLEQUIN_ACTIONS = {
     "results_viewer.view_cell": Action(
         target=ResultsTable, action="view_cell", description="View Cell"
     ),
+    "results_viewer.edit_cell": Action(
+        target=ResultsTable, action="edit_cell", description="Edit Cell"
+    ),
     "results_viewer.select_cursor": Action(target=ResultsTable, action="select_cursor"),
     # Moving the cursor
     "results_viewer.cursor_up": Action(target=ResultsTable, action="cursor_up"),